LayoutTests/imported/w3c/web-platform-tests/selection/textcontrols/selectionchange.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test selectionchange events from text controls</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="stylesheet" href="/fonts/ahem.css">
<style>
  input,
  textarea {
    font: 16px/1 Ahem;
  }
</style>

<input id="input" width="200"><br>
<textarea id="textarea" width="200"></textarea>

<script>
  class SelectionChangeCollector {
    /**
     * @param {HTMLElement} target
     */
    constructor(target) {
      this.target = target;
      this.events = [];
      target.addEventListener("selectionchange", ev => {
        this.events.push(ev);
      });
    }
    clear() {
      this.events.length = 0;
    }
  }

  const data = {
    collectors: [
      new SelectionChangeCollector(input),
      new SelectionChangeCollector(input.cloneNode()),
      new SelectionChangeCollector(textarea),
      new SelectionChangeCollector(textarea.cloneNode(true)),
    ],
    async initialize() {
      for (const collector of this.collectors) {
        collector.target.value = "XXXXXXXXXXXXXXXXXXX";
        collector.target.blur();
        collector.target.setSelectionRange(0, 0);
      }
      await this.spin();
      for (const collector of this.collectors) {
        collector.clear();
      }
    },
    spin() {
      return new Promise(setTimeout);
    },
    async assert_empty_spin() {
      // firing selectionchange must be asynchronous
      for (const collector of this.collectors) {
        assert_equals(collector.events.length, 0);
      }
      await this.spin();
    }
  };

  for (const collector of data.collectors) {
    const target = collector.target;
    const name = `the ${!target.parentNode ? "disconnected " : ""}${target.localName} element`;

    promise_test(async () => {
      await data.initialize();

      target.value = "";
      target.setRangeText("foo");

      await data.assert_empty_spin();
      assert_equals(collector.events.length, 0);
    }, `Calling setRangeText() on empty ${name}`);
  }
</script>
