LayoutTests/imported/w3c/web-platform-tests/selection/textcontrols/selectionchange-expected.txt




FAIL Calling setRangeText() on empty the input element assert_equals: expected 0 but got 1
FAIL Calling setRangeText() on empty the disconnected input element assert_equals: expected 0 but got 1
FAIL Calling setRangeText() on empty the textarea element assert_equals: expected 0 but got 1
FAIL Calling setRangeText() on empty the disconnected textarea element assert_equals: expected 0 but got 1
